composing Java-based configurations
