+++ myapp/src/main/java/com/learn/configs/AppConfig.java
@@ -1,15 +1,13 @@
 package com.learn.configs;
 
 import com.learn.repositories.MyRepositoryImpl;
-import com.learn.seedworks.InspectRegisteredScopesPostProcessor;
 import com.learn.services.TheirService;
-import org.springframework.beans.factory.config.CustomScopeConfigurer;
 import org.springframework.context.annotation.*;
-import org.springframework.context.support.SimpleThreadScope;
 
 @Configuration
 @ComponentScan(basePackages = "com.learn")
 @PropertySource("classpath:application.properties")
+@Import(SpringContainerConfig.class)
 public class AppConfig {
     @Bean
     @Scope("prototype")
@@ -17,15 +15,4 @@ public class AppConfig {
         return new TheirService(new MyRepositoryImpl());
     }
 
-    @Bean
-    public static CustomScopeConfigurer getCustomScopeConfigurer(){
-        var customScopeConfigurer = new CustomScopeConfigurer();
-        customScopeConfigurer.addScope("thread", new SimpleThreadScope());
-        return customScopeConfigurer;
-    }
-
-    @Bean
-    public static InspectRegisteredScopesPostProcessor getInspectScopesPostProcessor(){
-        return new InspectRegisteredScopesPostProcessor();
-    }
 }
